chore(arch): finish entity move (AlertStateChange + PushToken) + ArchUnit guard
L-1 from `necesito-esto-title-sleepy-piglet`.

Move the remaining JPA entities of the alert/push modules to their
respective infrastructure layers:

- features/alert/AlertStateChange.kt → features/alert/infrastructure/adapter/output/AlertStateChange.kt
- features/push/PushToken.kt → features/push/infrastructure/adapter/output/PushToken.kt
  (PushPlatform enum moves with it)

The legacy package files now contain only `typealias` shims so existing
importers (repositories, listeners, adapters) keep compiling without a
sweeping rename across ~30 files. Same shim pattern Pablo introduced for
Alert and User in PR #141.

Add ArchUnit rule `migratedEntitiesShouldResideInInfrastructure` scoped
to the modules already migrated (alert, user, push) so we cannot regress.
The remaining legacy modules (catalog, sector, device, greenhouse, mqtt,
notification, command, auth.refresh) keep their existing layout for now;
each will be added to the rule's package list as it gets migrated.
